some scripts are migrated to pyqt5.
