+++ traub_2005/py/display_morphology.py
